+++ src/Resources/contao/config/config.php
@@ -6,13 +6,12 @@
  * @license LGPL-3.0-or-later
  */
 
-use HeimrichHannot\SyndicationTypeBundle\ContentElement\SyndicationElement;
 use HeimrichHannot\SyndicationTypeBundle\EventListener\Contao\CompileArticleListener;
+use HeimrichHannot\SyndicationTypeBundle\EventListener\Contao\InitializeSystemListener;
 use HeimrichHannot\SyndicationTypeBundle\EventListener\Contao\LoadDataContainerListener;
 use HeimrichHannot\SyndicationTypeBundle\EventListener\Contao\ParseTemplateListener;
 
 $GLOBALS['TL_HOOKS']['loadDataContainer'][] = [LoadDataContainerListener::class, '__invoke'];
 $GLOBALS['TL_HOOKS']['compileArticle'][] = [CompileArticleListener::class, '__invoke'];
 $GLOBALS['TL_HOOKS']['parseTemplate'][] = [ParseTemplateListener::class, '__invoke'];
-
-$GLOBALS['TL_CTE']['links'][SyndicationElement::TYPE] = SyndicationElement::class;
+$GLOBALS['TL_HOOKS']['initializeSystem'][] = [InitializeSystemListener::class, '__invoke'];