+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2021-06-22
+- added new syndication types: xing, linkedin
+- moved content element creation to InitilizeSystemListner
+
 ## [0.3.0] - 2021-06-14
 - added content element syndication type(#1)